comments to rule of 5 methods in Plan

+++ src/SelectionPolicy.cpp
@@ -74,7 +74,7 @@ const FacilityType& EconomySelection::selectFacility(const vector<FacilityType>&
         throw std::runtime_error("No facilities available for selection.");
     }
     //Iterate over all facilities to find the first facility in the ECONOMY category
-    for (int i = lastSelectedIndex + 1; i < facilitiesOptions.size(); ++i) {
+    for (size_t i = lastSelectedIndex + 1; i < facilitiesOptions.size(); ++i) {
         if (facilitiesOptions[i].getCategory() == FacilityCategory::ECONOMY) {
             lastSelectedIndex = i;
             return facilitiesOptions[i];
@@ -101,7 +101,7 @@ const FacilityType& SustainabilitySelection::selectFacility(const vector<Facilit
         throw std::runtime_error("No facilities available for selection.");
     }
     //Iterate over all facilities to find the first facility in the ENVIRONMENT category
-    for (int i = lastSelectedIndex + 1; i < facilitiesOptions.size(); ++i) {
+    for (size_t i = lastSelectedIndex + 1; i < facilitiesOptions.size(); ++i) {
         if (facilitiesOptions[i].getCategory() == FacilityCategory::ENVIRONMENT) {
             lastSelectedIndex = i;
             return facilitiesOptions[i];
